Add support for domain forwarding as early as possible
The code that adds support for domain forwarding needs to be executed as
*early* as possible so as to set the $_SERVER['HTTPS'] env var before
any other core or plugin code needs it.

In order to achieve this, the installed/active plugin directory should
be named a-woocommerce-dev-helper so that it's executed before all other
plugins. Otherwise with EDD for instance, a bunch of external files will
be loaded over HTTP and raise mixed content errors when forwarding over
HTTPS